requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
beautifulsoup4==4.14.3
    # via agenda-rouen (pyproject.toml)
boto3==1.42.70
    # via agenda-rouen (pyproject.toml)
botocore==1.42.70
    # via
    #   boto3
    #   s3transfer
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
cryptography==46.0.5
    # via google-auth
distro==1.9.0
    # via google-genai
google-auth==2.49.1
    # via google-genai
google-genai==1.68.0
    # via agenda-rouen (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   agenda-rouen (pyproject.toml)
    #   google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
lxml==6.0.2
    # via agenda-rouen (pyproject.toml)
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   agenda-rouen (pyproject.toml)
    #   google-genai
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.2.2
    # via agenda-rouen (pyproject.toml)
requests==2.32.5
    # via
    #   google-auth
    #   google-genai
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via google-genai
soupsieve==2.8.3
    # via beautifulsoup4
tenacity==9.1.4
    # via google-genai
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via
    #   botocore
    #   requests
websockets==16.0
    # via google-genai
